Pull in MSG schemas and services by default
